+++ PuntoDeVenta/Capa Entidad/CE_Productos.cs
@@ -12,9 +12,9 @@ namespace Capa_Entidad
         private string nombre;
         private int grupo;
         private string codigo;
-        private double precio;
+        private decimal precio;
         private bool activo;
-        private double cantidad;
+        private decimal cantidad;
         private string unidadMedida;
         private byte[] img;
         private string descripcion;
@@ -23,9 +23,9 @@ namespace Capa_Entidad
         public string Nombre { get => nombre; set => nombre = value; }
         public int Grupo { get => grupo; set => grupo = value; }
         public string Codigo { get => codigo; set => codigo = value; }
-        public double Precio { get => precio; set => precio = value; }
+        public decimal Precio { get => precio; set => precio = value; }
         public bool Activo { get => activo; set => activo = value; }
-        public double Cantidad { get => cantidad; set => cantidad = value; }
+        public decimal Cantidad { get => cantidad; set => cantidad = value; }
         public string UnidadMedida { get => unidadMedida; set => unidadMedida = value; }
         public byte[] Img { get => img; set => img = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
